move dashboard to views directory
